refactor(tests): consolidate service tests into single-file-per-service pattern
Migrated 95 service tests from 14 separate files into 3 consolidated files
using nested classes with shared base classes. This eliminates code
duplication and improves test organization.

Changes:
- Consolidated ChatService tests (18 tests) into ChatServiceTests.cs
- Consolidated NotificationService tests (41 tests) into NotificationServiceTests.cs
- Consolidated ShareService tests (36 tests) into ShareServiceTests.cs
- Updated CLAUDE.md to reflect new test organization pattern
- All 125 tests passing
