--- BookSharingApp.Tests/Services/UpdateShareDueDateAsyncTests.cs
@@ -1,192 +1,0 @@
-using BookSharingApp.Common;
-using BookSharingApp.Services;
-using BookSharingApp.Tests.Helpers;
-using FluentAssertions;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Logging;
-using Moq;
-
-namespace BookSharingApp.Tests.Services
-{
-    public class UpdateShareDueDateAsyncTests : IDisposable
-    {
-        private readonly Mock<ILogger<ShareService>> _loggerMock;
-        private readonly Mock<INotificationService> _notificationServiceMock;
-
-        public UpdateShareDueDateAsyncTests()
-        {
-            _loggerMock = new Mock<ILogger<ShareService>>();
-            _notificationServiceMock = new Mock<INotificationService>();
-        }
-
-        public void Dispose()
-        {
-            // Cleanup if needed
-        }
-
-        [Fact]
-        public async Task UpdateShareDueDateAsync_ByLender_UpdatesReturnDateAndCreatesNotification()
-        {
-            // Arrange
-            using var context = DbContextHelper.CreateInMemoryContext();
-            var shareService = new ShareService(context, _loggerMock.Object, _notificationServiceMock.Object);
-
-            var lender = TestDataBuilder.CreateUser(id: "lender-1", firstName: "John", lastName: "Doe");
-            var borrower = TestDataBuilder.CreateUser(id: "borrower-1");
-            var book = TestDataBuilder.CreateBook(title: "The Great Gatsby");
-
-            context.Users.AddRange(lender, borrower);
-            context.Books.Add(book);
-            await context.SaveChangesAsync();
-
-            var userBook = TestDataBuilder.CreateUserBook(
-                userId: lender.Id,
-                bookId: book.Id,
-                user: lender,
-                book: book
-            );
-            context.UserBooks.Add(userBook);
-            await context.SaveChangesAsync();
-
-            var share = TestDataBuilder.CreateShare(
-                userBookId: userBook.Id,
-                borrower: borrower.Id,
-                status: ShareStatus.Ready,
-                userBook: userBook,
-                borrowerUser: borrower
-            );
-            context.Shares.Add(share);
-            await context.SaveChangesAsync();
-
-            var newDueDate = DateTime.UtcNow.AddDays(14);
-
-            // Act
-            await shareService.UpdateShareDueDateAsync(share.Id, newDueDate, lender.Id);
-
-            // Assert
-            var updatedShare = await context.Shares.FindAsync(share.Id);
-            updatedShare.Should().NotBeNull();
-            updatedShare!.ReturnDate.Should().BeCloseTo(newDueDate, TimeSpan.FromSeconds(1));
-
-            _notificationServiceMock.Verify(
-                x => x.CreateShareNotificationAsync(
-                    share.Id,
-                    NotificationType.ShareDueDateChanged,
-                    It.Is<string>(msg => msg.Contains("Return date") && msg.Contains("The Great Gatsby")),
-                    lender.Id
-                ),
-                Times.Once
-            );
-        }
-
-        [Fact]
-        public async Task UpdateShareDueDateAsync_ByBorrower_ThrowsUnauthorizedAccessException()
-        {
-            // Arrange
-            using var context = DbContextHelper.CreateInMemoryContext();
-            var shareService = new ShareService(context, _loggerMock.Object, _notificationServiceMock.Object);
-
-            var lender = TestDataBuilder.CreateUser(id: "lender-1");
-            var borrower = TestDataBuilder.CreateUser(id: "borrower-1");
-            var book = TestDataBuilder.CreateBook();
-
-            context.Users.AddRange(lender, borrower);
-            context.Books.Add(book);
-            await context.SaveChangesAsync();
-
-            var userBook = TestDataBuilder.CreateUserBook(
-                userId: lender.Id,
-                bookId: book.Id,
-                user: lender,
-                book: book
-            );
-            context.UserBooks.Add(userBook);
-            await context.SaveChangesAsync();
-
-            var share = TestDataBuilder.CreateShare(
-                userBookId: userBook.Id,
-                borrower: borrower.Id,
-                status: ShareStatus.PickedUp,
-                userBook: userBook,
-                borrowerUser: borrower
-            );
-            context.Shares.Add(share);
-            await context.SaveChangesAsync();
-
-            var newDueDate = DateTime.UtcNow.AddDays(7);
-
-            // Act
-            var act = async () => await shareService.UpdateShareDueDateAsync(share.Id, newDueDate, borrower.Id);
-
-            // Assert
-            await act.Should().ThrowAsync<UnauthorizedAccessException>()
-                .WithMessage("Only the lender can set the return date");
-        }
-
-        [Fact]
-        public async Task UpdateShareDueDateAsync_WhenShareNotFound_ThrowsInvalidOperationException()
-        {
-            // Arrange
-            using var context = DbContextHelper.CreateInMemoryContext();
-            var shareService = new ShareService(context, _loggerMock.Object, _notificationServiceMock.Object);
-
-            var lender = TestDataBuilder.CreateUser(id: "lender-1");
-            context.Users.Add(lender);
-            await context.SaveChangesAsync();
-
-            var newDueDate = DateTime.UtcNow.AddDays(14);
-
-            // Act
-            var act = async () => await shareService.UpdateShareDueDateAsync(999, newDueDate, lender.Id);
-
-            // Assert
-            await act.Should().ThrowAsync<InvalidOperationException>()
-                .WithMessage("Share not found");
-        }
-
-        [Fact]
-        public async Task UpdateShareDueDateAsync_ByUnauthorizedUser_ThrowsUnauthorizedAccessException()
-        {
-            // Arrange
-            using var context = DbContextHelper.CreateInMemoryContext();
-            var shareService = new ShareService(context, _loggerMock.Object, _notificationServiceMock.Object);
-
-            var lender = TestDataBuilder.CreateUser(id: "lender-1");
-            var borrower = TestDataBuilder.CreateUser(id: "borrower-1");
-            var unauthorizedUser = TestDataBuilder.CreateUser(id: "unauthorized-user");
-            var book = TestDataBuilder.CreateBook();
-
-            context.Users.AddRange(lender, borrower, unauthorizedUser);
-            context.Books.Add(book);
-            await context.SaveChangesAsync();
-
-            var userBook = TestDataBuilder.CreateUserBook(
-                userId: lender.Id,
-                bookId: book.Id,
-                user: lender,
-                book: book
-            );
-            context.UserBooks.Add(userBook);
-            await context.SaveChangesAsync();
-
-            var share = TestDataBuilder.CreateShare(
-                userBookId: userBook.Id,
-                borrower: borrower.Id,
-                status: ShareStatus.Ready,
-                userBook: userBook,
-                borrowerUser: borrower
-            );
-            context.Shares.Add(share);
-            await context.SaveChangesAsync();
-
-            var newDueDate = DateTime.UtcNow.AddDays(14);
-
-            // Act
-            var act = async () => await shareService.UpdateShareDueDateAsync(share.Id, newDueDate, unauthorizedUser.Id);
-
-            // Assert
-            await act.Should().ThrowAsync<UnauthorizedAccessException>()
-                .WithMessage("Only the lender can set the return date");
-        }
-    }
-}
